bot/pipe_line_v1/requirements.txt
# Core dependencies
langchain==0.3.0
langchain-openai==0.2.0
openai==1.93.0
python-dotenv==1.1.1

# Data processing
pandas==2.2.3
numpy==1.26.0

# Clustering and ML
hdbscan==0.8.33
scikit-learn==1.5.0

# NLP
spacy==3.7.2

# Database
SQLAlchemy==2.0.42

# Utilities
requests==2.31.0
aiohttp==3.9.1
